updated dependencies to newest version, & fixed slow startup issue
