+++ build.gradle
@@ -12,33 +12,30 @@ sourceCompatibility = targetCompatibility = 1.8
 
 repositories {
     mavenCentral()
-    maven {
-        name 'm2-dv8tion'
-        url 'https://m2.dv8tion.net/releases'
-    }
 }
 
 dependencies {
     /**
      * JCenter imports
      */
-    //https://github.com/DV8FromTheWorld/JDA
-    compile "net.dv8tion:JDA:4.4.0_350"
+    //https://github.com/discord-jda/jda
+    implementation("net.dv8tion:JDA:5.0.0-beta.18") {
+        exclude module: 'opus-java'
+    }
 
     /**
      * Maven Central imports
      */
-    testCompile group: 'junit', name: 'junit', version: '4.12'
     // https://mvnrepository.com/artifact/com.google.code.gson/gson
-    compile group: 'com.google.code.gson', name: 'gson', version: '2.8.6'
+    implementation(group: 'com.google.code.gson', name: 'gson', version: '2.10.1')
     // https://mvnrepository.com/artifact/com.googlecode.json-simple/json-simple
-    compile group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1.1'
+    implementation(group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1.1')
     // https://mvnrepository.com/artifact/com.vdurmont/emoji-java
-    compile group: 'com.vdurmont', name: 'emoji-java', version: '5.1.1'
+    implementation(group: 'com.vdurmont', name: 'emoji-java', version: '5.1.1')
     // https://mvnrepository.com/artifact/it.sauronsoftware.cron4j/cron4j
-    compile group: 'it.sauronsoftware.cron4j', name: 'cron4j', version: '2.2.5'
+    implementation(group: 'it.sauronsoftware.cron4j', name: 'cron4j', version: '2.2.5')
     // https://mvnrepository.com/artifact/org.jboss.resteasy/resteasy-client
-    compile group: 'org.jboss.resteasy', name: 'resteasy-client', version: '5.0.1.Final'
+    implementation(group: 'org.jboss.resteasy', name: 'resteasy-client', version: '6.2.6.Final')
 
 }
